sql/tests/requirements.txt
pytest
pyyaml
pandasql
mlflow
sklearn
matplotlib
numpy==1.19.5
ctgan==0.2.2.dev0
opacus==0.11.0
torch==1.7.1
gym
flake8
diffprivlib
graphviz
